gs-hack-education/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
requests = "^2.26.0"
newsapi-python = "^0.2.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
